Refactor CompactCard
